pyproject.toml
[build-system]
requires = ["setuptools>=61.0"]
build-backend = "setuptools.build_meta"

[project]
name = "zettelkasten"
version = "0.1.0"
description = "A tool for processing audio notes into a knowledge base"
readme = "README.md"
requires-python = ">=3.10"
authors = [
    {name = "Zack Bodnar"}
]
dependencies = [
    "openai>=1.3.0",
    "openai-whisper @ git+https://github.com/openai/whisper.git",
    "deepgram-sdk>=3.10.0",
    "pyyaml>=6.0.1",
    "tiktoken>=0.5.2",
    "chromadb>=0.4.22",
    "neo4j>=5.14.1",
    "python-dotenv>=1.0.0",
    "pydub>=0.25.1",
    "soundfile>=0.12.1",
    "numpy>=1.24.3",
    "scipy>=1.11.3",
    "torch>=2.0.0",
    "transformers>=4.30.0",
    "nltk>=3.8.1",
    "tqdm>=4.65.0",
    "ffmpeg-python>=0.2.0"
]

[project.optional-dependencies]
dev = [
    "pytest>=7.4.3",
    "pytest-asyncio>=0.21.1"
] 